Replaced Apollo, Finished VM2, Fixed some bugs, Error Handling, Add gql-playground module, some refactoring

+++ modules/graphql/src/getDefaultResolvers/MutationBuilder/resolveDeleteMutation.ts
@@ -1,15 +1,23 @@
-import { IParsedResolveInfo } from "../types";
-import { IMutationBuildObject, IMutationInputObject } from "./types";
+import { IMutationBuildObject } from "./types";
 import { ReturnIdHandler } from "../../resolverTransactions/ReturnIdHandler";
 import { IMutationViewMeta, IDefaultResolverMeta } from "../../RuntimeInterfaces";
+import { ResolveTree } from "graphql-parse-resolve-info";
 
 export default function resolveDeleteMutation(
   defaultResolverMeta: IDefaultResolverMeta,
-  query: IParsedResolveInfo<IMutationInputObject>,
+  query: ResolveTree,
   mutation: IMutationViewMeta,
   returnIdHandler: ReturnIdHandler
 ): IMutationBuildObject {
-  const entityId: string = returnIdHandler.getReturnId(query.args.input.id);
+  if (query.args.input == null) {
+    throw new Error("Invalid input.");
+  }
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const input: any = query.args.input;
+  if (input.id == null) {
+    throw new Error("Invalid input. Missing id.");
+  }
+  const entityId: string = returnIdHandler.getReturnId(input.id);
   return {
     sql: `DELETE FROM "${defaultResolverMeta.viewsSchemaName}"."${mutation.viewName}" WHERE id = $1;`,
     values: [entityId],
